Query: Replace the "Reload full page enabled" modal with a snackbar notice
